Reconstruct the hierarchy of perftop, seperate rca and metrics analysis
