+++ lib/bin.js
@@ -19,12 +19,15 @@ var path = require('path');
 var process = require('process');
 var util = require('util');
 
-var clusterOverviewJSON = require('./dashboards/ClusterOverview.json');
-var clusterNetworkMemoryAnalysisJSON = require('./dashboards/ClusterNetworkMemoryAnalysis.json');
-var clusterThreadAnalysisJSON = require('./dashboards/ClusterThreadAnalysis.json');
-var nodeAnalysisJSON = require('./dashboards/NodeAnalysis.json');
-var temperatureAnalysisJSON = require('./dashboards/TemperatureAnalysis.json');
-var graphGenerator = require('./perf-top/generate-graphs.js');
+var clusterOverviewJSON = require('./dashboards/metrics/ClusterOverview.json');
+var clusterNetworkMemoryAnalysisJSON = require('./dashboards/metrics/ClusterNetworkMemoryAnalysis.json');
+var clusterThreadAnalysisJSON = require('./dashboards/metrics/ClusterThreadAnalysis.json');
+var nodeAnalysisJSON = require('./dashboards/metrics/NodeAnalysis.json');
+
+var temperatureAnalysisJSON = require('./dashboards/rca/TemperatureAnalysis.json');
+
+var metricsGraphGenerator = require('./perf-top/metrics/generate-graphs.js');
+var rcaGraphGenerator = require('./perf-top/rca/generate-graphs')
 
 // Parse command line arguments
 var parser = new argumentParser({
@@ -91,7 +94,7 @@ console.error = function (msg) {
 };  
 
 if (args.mode === 'rca') {
-  graphGenerator.initAndStartRCA(jsonData);
+  rcaGraphGenerator.initAndStart(jsonData);
 } else {
-  graphGenerator.initAndStart(jsonData);
+  metricsGraphGenerator.initAndStart(jsonData);
 }
