+++ lib/dashboards/metrics/ClusterThreadAnalysis.json
