+++ lib/dashboards/metrics/NodeAnalysis.json
