+++ lib/perf-top/metric-graphs.js
