+++ lib/perf-top/metrics/generate-graphs.js
@@ -17,12 +17,9 @@ require('console-stamp')(console, '[HH:MM:ss.l]');
 
 var dataGenerator = require('./util/generate-data.js');
 var metricBar = require('./util/metric-bar.js');
-var metricGraphs = require('./util/metric-graphs.js');
+var metricGraphs = require('../metric-graphs');
 var metricLine = require('./util/metric-line.js');
 var metricTable = require('./util/metric-table.js');
-var metricDonutRCA = require('./util/metric-donut-rca.js');
-var metricLineRCA = require('./util/metric-line-rca.js');
-var metricTableRCA = require('./util/metric-table-rca.js');
 
 /**
  * Initialize all the graph objects and generate them.
@@ -55,35 +52,4 @@ function initAndStart (jsonData) {
   });
 }
 
-/**
- * Initialize all the graph objects and generate RCA analysis.
- *
- * @param {object} jsonData - hashmap of dashboard configuration.
- */
-function initAndStartRCA (jsonData) {
-  var graphs = new metricGraphs.metricGraphs();
-  dataGenerator.getMetricUnits(jsonData.endpoint, function () { 
-    var queryParams = jsonData.queryParams;
-    for (var i = 0; i < jsonData.graphs.length; i++) {
-      var graphConfig = jsonData.graphs[i];
-      var graphType = graphConfig.graphType;
-      if ((graphType === 'donuts')) {
-        graph = new metricDonutRCA.metricDonutRCA(jsonData.endpoint, jsonData.gridOptions, queryParams,
-          graphConfig.options, graphConfig.dimension, graphConfig.graphParams, graphs.screen);
-      } else if ((graphType === 'lines')) {
-        graph = new metricLineRCA.metricLineRCA(jsonData.endpoint, jsonData.gridOptions, queryParams,
-          graphConfig.options, graphConfig.dimension, graphConfig.graphParams, graphs.screen);
-      } else if (graphType === 'tables') {
-        graph = new metricTableRCA.metricTableRCA(jsonData.endpoint, jsonData.gridOptions, queryParams,
-          graphConfig.options, graphConfig.dimension, graphConfig.graphParams, graphs.screen);
-      }
-      graphs.allGraphs.push(graph);
-    }
-    // Generate graph on screen
-    graphs.resizeGraphsToScreen();
-    graphs.start();
-  });
-}
-
 module.exports.initAndStart = initAndStart;
-module.exports.initAndStartRCA = initAndStartRCA;
