+++ lib/perf-top/metrics/util/metric-bar.js
