+++ lib/perf-top/metrics/util/metric-line.js
