+++ lib/perf-top/metrics/util/metric-table.js
