+++ lib/perf-top/metrics/util/validate-query-params.js
