+++ lib/perf-top/rca/util/metric-donut.js
@@ -14,7 +14,6 @@
  */
 
 var contrib = require('blessed-contrib');
-var dataGenerator = require('./generate-data.js');
 
 /**
  * Creates the donut graph, which aggregates data on all nodes unless 'nodeName' defined.
@@ -26,14 +25,16 @@ var dataGenerator = require('./generate-data.js');
  * @param {object} options - options for contrib.donut() object.
  * @param {string} dimension - dimension of RCA to be analyzed.
  * @param {string} graphParams - the parameter to be rendered in the graph.
+ * @param {object} dataGenerator - the data generator for the rca dimension
  * @param {object} screen - blessed.screen() object.
  */
-function metricDonutRCA (endpoint, gridOptions, queryParams, options, dimension, graphParams, screen) {
+function metricDonut (endpoint, gridOptions, queryParams, options, dimension, graphParams, dataGenerator, screen) {
     this.endpoint = endpoint;
     this.name = queryParams.name;
     this.local = queryParams.local;
     this.dimension = dimension;
     this.graphParams = graphParams;
+    this.dataGenerator = dataGenerator;
     this.refreshInterval = (options.refreshInterval > 5000) ? options.refreshInterval : 5000;
 
     var grid = new contrib.grid({ rows: gridOptions.rows, cols: gridOptions.cols, screen: screen });
@@ -46,7 +47,7 @@ function metricDonutRCA (endpoint, gridOptions, queryParams, options, dimension,
 /**
  * Wrapper to attach the donut object to screen.
  */
-metricDonutRCA.prototype.emit = function () {
+metricDonut.prototype.emit = function () {
   this.donut.emit('attach');
 };
 
@@ -56,8 +57,8 @@ metricDonutRCA.prototype.emit = function () {
  * @param {object} donut - metricDonut() object.
  * @param {object} screen - blessed.screen() object.
  */
-metricDonutRCA.prototype.generateGraph = function (donut, screen) {
-    generateMetricDonutData(donut, function (donutData) {
+metricDonut.prototype.generateGraph = function (donut, screen) {
+  generateMetricDonutData(donut, function (donutData) { ///problem
     if (Object.keys(donutData).length !== 0) {
       donut.donut.setData(donutData);
       screen.render();
@@ -81,8 +82,9 @@ metricDonutRCA.prototype.generateGraph = function (donut, screen) {
  *                                               ]                                              
  */
 function generateMetricDonutData (metricDonut, callback) {
-  dataGenerator.getMetricDataRCA(metricDonut.endpoint, metricDonut.name, metricDonut.local, metricDonut.dimension, metricDonut.graphParams, "donuts", function (metricData) {
-    dataGenerator.removeStaleDataRCA(metricData, metricDonut);
+  var dataGenerator = metricDonut.dataGenerator;
+  dataGenerator.getMetricData(metricDonut.endpoint, metricDonut.name, metricDonut.local, metricDonut.dimension, metricDonut.graphParams, "donuts", function (metricData) {
+    dataGenerator.removeStaleData(metricData, metricDonut);
     var color = ["red", "yellow", "green", "blue"];
     var donutData = [];
     var count = metricData.data.reduce(function(a, b) {
@@ -98,4 +100,4 @@ function generateMetricDonutData (metricDonut, callback) {
   });
 }
 
-module.exports.metricDonutRCA = metricDonutRCA;
+module.exports.metricDonut = metricDonut;
