+++ lib/perf-top/rca/util/metric-line.js
@@ -14,7 +14,6 @@
  */
 
 var contrib = require('blessed-contrib');
-var dataGenerator = require('./generate-data.js');
 
 /**
  * Creates the line graph, which represents data on per node level.
@@ -26,15 +25,17 @@ var dataGenerator = require('./generate-data.js');
  * @param {object} options - options for contrib.line() object.
  * @param {string} dimension - dimension of RCA to be analyzed.
  * @param {string} graphParams - the parameter to be rendered in the graph.
+ * @param {object} dataGenerator - the data generator for the rca dimension
  * @param {object} screen - blessed.screen() object.
  */
-function metricLineRCA (endpoint, gridOptions, queryParams, options, dimension, graphParams, screen) {
+function metricLine (endpoint, gridOptions, queryParams, options, dimension, graphParams, dataGenerator, screen) {
   this.endpoint = endpoint;
   this.name = queryParams.name;
   this.local = queryParams.local; 
   this.dimension = dimension;
   this.graphParams = graphParams;
   this.lines = {};
+  this.dataGenerator = dataGenerator;
   this.refreshInterval = (options.refreshInterval > 5000) ? options.refreshInterval : 5000;
 
   var grid = new contrib.grid({ rows: gridOptions.rows, cols: gridOptions.cols, screen: screen });
@@ -51,7 +52,7 @@ function metricLineRCA (endpoint, gridOptions, queryParams, options, dimension, 
 /**
  * Wrapper to attach the line object to screen.
  */
-metricLineRCA.prototype.emit = function () {
+metricLine.prototype.emit = function () {
   this.line.emit('attach');
 };
 
@@ -61,7 +62,7 @@ metricLineRCA.prototype.emit = function () {
  * @param {object} line - metricLine() object.
  * @param {object} screen - blessed.screen() object.
  */
-metricLineRCA.prototype.generateGraph = function (line, screen) {
+metricLine.prototype.generateGraph = function (line, screen) {
   generateAndUpdateLineData(line);
   screen.render();
 };
@@ -97,7 +98,8 @@ function updateLineData (lineData, newData) {
  * @param {object} metricLine - metricLine() object.
  */
 function generateAndUpdateLineData (metricLine) {
-  dataGenerator.getMetricDataRCA(metricLine.endpoint, metricLine.name, metricLine.local, metricLine.dimension, metricLine.graphParams, "lines",
+  var dataGenerator = metricLine.dataGenerator;
+  dataGenerator.getMetricData(metricLine.endpoint, metricLine.name, metricLine.local, metricLine.dimension, metricLine.graphParams, "lines",
     function (metricData) { 
       if (metricData.fields === null || metricData.fields.length === 0) {
         console.error(`Metric was not found for request with queryParams:\n
@@ -106,7 +108,7 @@ function generateAndUpdateLineData (metricLine) {
         local: ${metricLine.local}\n`);
         return;
       }   
-      dataGenerator.removeStaleDataRCA(metricData, metricLine);
+      dataGenerator.removeStaleData(metricData, metricLine);
       var field = metricLine.dimension + " - " + metricLine.graphParams;
       if (!Object.keys(metricLine.lines).includes(field)) {
         metricLine.lines[field] = { 
@@ -142,4 +144,4 @@ function randomColor (colorChoices) {
   }
 }
 
-module.exports.metricLineRCA = metricLineRCA;
+module.exports.metricLine = metricLine;
