+++ lib/perf-top/rca/util/metric-table.js
@@ -14,7 +14,6 @@
  */
 
 var contrib = require('blessed-contrib');
-var dataGenerator = require('./generate-data.js');
 
 /**
  * Creates a table graph, which represents data per node-level.
@@ -26,18 +25,18 @@ var dataGenerator = require('./generate-data.js');
  * @param {object} options - options for contrib.table() object.
  * @param {string} dimension - dimension of RCA to be analyzed.
  * @param {string} graphParams - the parameter to be rendered in the graph.
+ * @param {object} dataGenerator - the data generator for the rca dimension
  * @param {object} screen - blessed.screen() object.
  */
-function metricTableRCA (endpoint, gridOptions, queryParams, options, dimension, graphParams, screen) {
+function metricTable (endpoint, gridOptions, queryParams, options, dimension, graphParams, dataGenerator, screen) {
   this.endpoint = endpoint;
   this.name = queryParams.name;
   this.local = queryParams.local;
   this.dimension = dimension;
   this.graphParams = graphParams;
   this.labels = (options.columns).split(',');
-
+  this.dataGenerator = dataGenerator;
   options.columnWidth = this.labels.map(label => label.length + 10);
-
   this.refreshInterval = (options.refreshInterval > 5000) ? options.refreshInterval : 5000;
 
   var grid = new contrib.grid({ rows: gridOptions.rows, cols: gridOptions.cols, screen: screen });
@@ -50,7 +49,7 @@ function metricTableRCA (endpoint, gridOptions, queryParams, options, dimension,
 /**
  * Wrapper to attach the table object to screen.
  */
-metricTableRCA.prototype.emit = function () {
+metricTable.prototype.emit = function () {
   this.table.emit('attach');
 };
 
@@ -60,7 +59,7 @@ metricTableRCA.prototype.emit = function () {
  * @param {object} table - metricTable object.
  * @param {object} screen - blessed.screen() object.
  */
-metricTableRCA.prototype.generateGraph = function (table, screen) {
+metricTable.prototype.generateGraph = function (table, screen) {
   generateMetricTableData(table, function (tableData) {
     if (Object.keys(tableData).length !== 0) {
       table.table.setData(tableData);
@@ -87,12 +86,13 @@ metricTableRCA.prototype.generateGraph = function (table, screen) {
  *                                           }
  */
 function generateMetricTableData (metricTable, callback) {
-  dataGenerator.getMetricDataRCA(metricTable.endpoint, metricTable.name, metricTable.local, metricTable.dimension, metricTable.graphParams, "tables",
+  var dataGenerator = metricTable.dataGenerator;
+  dataGenerator.getMetricData(metricTable.endpoint, metricTable.name, metricTable.local, metricTable.dimension, metricTable.graphParams, "tables",
     function (metricData) {
       if (metricData.fields === null || metricData.fields.length === 0) {
         callback({});
       }
-      dataGenerator.removeStaleDataRCA(metricData, metricTable);
+      dataGenerator.removeStaleData(metricData, metricTable);
       if (metricData.fields.length === 0) {
         callback({});
       } else {
@@ -108,4 +108,4 @@ function generateMetricTableData (metricTable, callback) {
     });
 }
 
-module.exports.metricTableRCA = metricTableRCA;
+module.exports.metricTable = metricTable;
